Correction problème de wget
(cherry picked from commit c7c553562d723f4ba789eca8468546a82850799d)
