+++ .github/dependabot.yml
@@ -15,7 +15,7 @@ updates:
     schedule:
       interval: "daily"
   - package-ecosystem: "npm" # See documentation for possible values
-    directory: "/data-management/tools/install-odb" # Location of package manifests
+    directory: "/tools/cli-tools" # Location of package manifests
     schedule:
       interval: "daily"
   - package-ecosystem: "npm" # See documentation for possible values
