+++ data-management/sonar-project.properties
@@ -1,5 +1,5 @@
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.sources=src,tools/install-odb/package.json,tools/install-odb/src,tools/model-generator/src,tools/model-generator/package.json
+sonar.sources=src,tools/model-generator/src,tools/model-generator/package.json
 sonar.tests=test
 
 sonar.eslint.reportPaths=tools/model-generator/dist/reports/lint/eslint-report.json
@@ -8,4 +8,4 @@ sonar.eslint.reportPaths=tools/model-generator/dist/reports/lint/eslint-report.j
 # sonar.exclusions=src/model/sqlite/**,src/model/mysql/**,src/model/pgsql/**,src/connector/**
 
 # cppcheck report for Sonar
-sonar.externalIssuesReportPaths=reports/cppcheck-sonarqube-report.json,reports/clang-tidy-sonarqube-report.json,reports/audit-install-odb-report.json,reports/audit-model-generator-report.json
+sonar.externalIssuesReportPaths=reports/cppcheck-sonarqube-report.json,reports/clang-tidy-sonarqube-report.json,reports/audit-model-generator-report.json
