--- data-management/tools/install-odb/.gitignore
@@ -1,3 +1,0 @@
-node_modules
-*.zip
-audit-dependency-install-odb-report.json
