+++ sonar-project.properties
@@ -7,7 +7,7 @@ sonar.projectName=hangman-game
 sonar.projectVersion=1.0.0
 
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.modules=client,server,data-management
+sonar.modules=client,server,data-management,tools
 # Encoding of the source code. Default is default system encoding
 sonar.sourceEncoding=UTF-8
