+++ tools/cli-tools/.gitignore
@@ -1,0 +1,3 @@
+node_modules
+*.zip
+audit-dependency-cli-tools-report.json
