+++ tools/sonar-project.properties
@@ -1,0 +1,4 @@
+sonar.sources=cli-tools/package.json,cli-tools/src
+
+# external report for Sonar
+sonar.externalIssuesReportPaths=reports/audit-cli-tools-report.json
